requirements.txt
astroid==2.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-cleanup==4.0.0
django-jet==1.0.8
django-parler==2.0
django-rosetta==0.9.3
django-widget-tweaks==1.4.5
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
pkg-resources==0.0.0
polib==1.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pylint==2.4.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
